micro/requirements.txt
fastapi
uvicorn
pdf2image
Pillow
python-multipart
python-docx
mammoth
